+++ crates/tls/Cargo.toml
@@ -26,6 +26,7 @@ authors.workspace = true
 name = "forge_tls"
 
 [dependencies]
+rustls = { workspace = true }
 serde = { features = ["derive"], workspace = true }
 serde_json = { workspace = true }
 tonic = { workspace = true }
