+++ crates/tls/src/lib.rs
@@ -21,3 +21,4 @@
 
 pub mod client_config;
 pub mod default;
+pub mod dummy_tls_verifier;
